clearing out app.tsx
